// sage-hr/src/emp/dto/SageHREmployee.dto.ts
export class SageHREmployee {
  id: number;
  email: string;
  first_name: string;
  last_name: string;
  picture_url?: string;
  employment_start_date?: string;
  date_of_birth?: string;
  position?: string;
  position_id?: number;
  work_phone?: string;
  mobile_phone?: string;
  home_phone?: string;
  employment_status?: string;
  termination_date?: string;
}